Move metrics behind a feature flag
